+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## 0.4.11 
 
+* Added `HttpReuqest::extract_xxx()`, type safe path/query information extractor.
+
 * Fix long client urls #129
 
 * Fix panic on invalid URL characters #130